+++ backend/src/main/java/com/backend/domain/user/dto/UserResponseDto.java
@@ -22,6 +22,6 @@ public class UserResponseDto {
     private String zipCode;
     private String phone;
     private String username;
-    private int restCash;
+    private Long restCash;
 
 }
